Correct FullMergePath which also works when duplicate names are used.
