+++ pkg/unittest/test_suite.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/helm-unittest/helm-unittest/pkg/unittest/results"
 	"github.com/helm-unittest/helm-unittest/pkg/unittest/snapshot"
-	"github.com/mitchellh/copystructure"
 	"gopkg.in/yaml.v3"
 	v3loader "helm.sh/helm/v3/pkg/chart/loader"
 	v3util "helm.sh/helm/v3/pkg/chartutil"
@@ -60,21 +59,6 @@ func createTestSuite(suiteFilePath string, chartRoute string, content string, st
 	suite.Values = append(suite.Values, valueFilesSet...)
 
 	return &suite, nil
-}
-
-func copySet(setValues map[string]interface{}) map[string]interface{} {
-	copiedSet, err := copystructure.Copy(setValues)
-	if err != nil {
-		panic(err)
-	}
-
-	copiedSetValues := copiedSet.(map[string]interface{})
-	// if we have an empty map, make sure it is initialized
-	if copiedSetValues == nil {
-		copiedSetValues = make(map[string]interface{})
-	}
-
-	return copiedSetValues
 }
 
 // RenderTestSuiteFiles renders a helm suite of test files and returns their TestSuites
