+++ test/data/v3/with-samenamesubsubcharts/README.md
@@ -1,0 +1,3 @@
+# example-chart
+
+This repository contains a set-up to reproduce the issue described in <https://github.com/quintush/helm-unittest/issues/111>.
