+++ test/data/v3/with-samenamesubsubcharts/charts/with-subsubchartssub/charts/with-subsubchartssubsub/Chart.yaml
@@ -1,0 +1,3 @@
+apiVersion: v2
+name: with-samenamesubsubcharts
+version: 0.3.0
