+++ test/data/v3/with-samenamesubsubcharts/charts/with-subsubchartssub/charts/with-subsubchartssubsub/templates/deployment.yaml
@@ -1,0 +1,11 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: example-sub-subchart
+spec:
+  template:
+    metadata:
+      annotations:
+      {{- with .Values.global.app.annotations }}
+        {{- toYaml . | trim | nindent 8 }}
+      {{- end }}
