+++ test/data/v3/with-samenamesubsubcharts/charts/with-subsubchartssub/charts/with-subsubchartssubsub/tests/deployment_test.yaml
@@ -1,0 +1,12 @@
+templates:
+  - deployment.yaml
+tests:
+  - it: should only add annotation `foo` with value `bar` and base localKey
+    set:
+      global.app.annotations:
+        foo: bar
+    asserts:
+      - equal:
+          path: spec.template.metadata.annotations
+          value:
+            foo: bar
