+++ test/data/v3/with-samenamesubsubcharts/charts/with-subsubchartssub/templates/deployment.yaml
@@ -1,0 +1,9 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: example-subchart
+spec:
+  template:
+    metadata:
+      annotations:
+        foo: sub
