+++ test/data/v3/with-samenamesubsubcharts/charts/with-subsubchartssub/values.yaml
@@ -1,0 +1,4 @@
+
+local:
+  annotations:
+    localKey: "localValue"
