+++ test/data/v3/with-samenamesubsubcharts/values.yaml
@@ -1,0 +1,3 @@
+global:
+  app:
+    annotations: {}
